requirements.txt
amqp==5.0.9
anyjson==0.3.3
asgiref==3.5.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.3
celery-progress==0.1.2
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cssselect==1.1.0
d==0.2.2
Deprecated==1.2.13
dj-database-url==0.5.0
Django==4.0.2
django-celery==3.3.1
django-celery-results==2.2.0
djangorestframework==3.13.1
Flask==2.0.3
idna==3.3
importlib-metadata==4.11.1
itsdangerous==2.1.1
Jinja2==3.0.3
joblib==1.1.0
kombu==5.2.3
lxml==4.8.0
Markdown==3.3.6
MarkupSafe==2.1.1
nltk==3.7
packaging==21.3
prompt-toolkit==3.0.28
Pygments==2.11.2
pyparsing==3.0.7
pyquery==1.4.3
pytz==2021.3
redis==4.1.4
regex==2022.3.2
requests==2.27.1
six @ file:///AppleInternal/BuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-103/six-1.15.0-py2.py3-none-any.whl
sqlparse==0.4.2
tqdm==4.63.0
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.3
whitenoise==6.0.0
wrapt==1.13.3
zipp==3.7.0
